requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.3
colorama==0.4.3
Django==3.0.7
django-filter==2.3.0
django-filters==0.2.1
djangorestframework==3.11.0
drf-nested-routers==0.91
flake8==3.8.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
